fixed error with empty vector. compiles and runs with two warnings

+++ Euchre/Euchre/Player.cpp
@@ -135,7 +135,7 @@ void Player::playCard(Card &card)
 	Card *a = &card;
 	Card *b;
 
-	for (unsigned int i = 0; i < m_hand.size(); i++) { //for every card in hand
+	for (int i = 0; i < (int)(m_hand.size()); i++) { //for every card in hand
 		b = &m_hand[i]; //get card at current index
 		if (b == a) { //if card matches
 			m_hand.erase(m_hand.begin() + i); //remove it by index
